unset TZ on Windows
TZ gets set when opening a cygwin terminal.

What I'm oberving is strange.. when TZ is set, even if it's set to the same
thing as the system time zone, it seems to result in files showing with
different mtimes than when TZ is not set.

Having TZ set also prevents seeing the real system timezone, so let's kill
it.