remove testenv threading from test suite
Simplify the test suite now that I have a way to setEnv on Windows by not
threading an environement through it.

Incidentially this fixed a failure on Windows, observed when running the
test suite in cygwin terminal, where there is apparently an env var named
"", and it tried to set that when propigating the environment, which failed.

But mostly, it makes the code a lot less horrible.

This commit was sponsored by Rémi Vanicat.